requirements.txt

telebot
python-dotenv
requests
python-telegram-bot
google.generativeai

